server/requirements.txt
fastapi==0.115.6
langchain_core==0.3.29
langchain_groq==0.2.2
pydantic==2.10.4
python-dotenv==1.0.1
uvicorn==0.27.1
